refactor(core): promote pkg/cli to internal/core/cli (no behavior change)
