+++ cmd/databricks-claude/desktop_config.go
@@ -16,8 +16,8 @@ import (
 	"time"
 
 	"github.com/IceRhymers/databricks-agents/internal/cmd"
+	"github.com/IceRhymers/databricks-agents/internal/core/cli"
 	"github.com/IceRhymers/databricks-agents/pkg/authcheck"
-	"github.com/IceRhymers/databricks-agents/pkg/cli"
 	"github.com/IceRhymers/databricks-agents/pkg/mdmprofile"
 	"github.com/IceRhymers/databricks-agents/pkg/modeldiscovery"
 )
