+++ cmd/databricks-claude/main.go
@@ -15,9 +15,9 @@ import (
 	"time"
 
 	"github.com/IceRhymers/databricks-agents/internal/cmd"
+	"github.com/IceRhymers/databricks-agents/internal/core/cli"
 	"github.com/IceRhymers/databricks-agents/internal/core/refcount"
 	"github.com/IceRhymers/databricks-agents/pkg/authcheck"
-	"github.com/IceRhymers/databricks-agents/pkg/cli"
 	"github.com/IceRhymers/databricks-agents/pkg/completion"
 	"github.com/IceRhymers/databricks-agents/pkg/health"
 	"github.com/IceRhymers/databricks-agents/pkg/mdmprofile"
