+++ cmd/databricks-claude/serve_install.go
@@ -10,8 +10,8 @@ import (
 	"time"
 
 	"github.com/IceRhymers/databricks-agents/internal/cmd"
+	"github.com/IceRhymers/databricks-agents/internal/core/cli"
 	"github.com/IceRhymers/databricks-agents/pkg/authcheck"
-	"github.com/IceRhymers/databricks-agents/pkg/cli"
 	"github.com/IceRhymers/databricks-agents/pkg/health"
 	"github.com/IceRhymers/databricks-agents/pkg/mdmprofile"
 )
