+++ cmd/databricks-claude/token.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/IceRhymers/databricks-agents/pkg/cli"
+	"github.com/IceRhymers/databricks-agents/internal/core/cli"
 	"github.com/IceRhymers/databricks-agents/pkg/tokencache"
 )
