+++ internal/core/cli/cli_test.go
